requirements/dev.txt
asgiref==3.5.2 ; python_version >= "3.10" and python_version < "4.0"
astroid==2.12.10 ; python_version >= "3.10" and python_version < "4.0"
attrs==22.1.0 ; python_version >= "3.10" and python_version < "4.0"
black==22.8.0 ; python_version >= "3.10" and python_version < "4.0"
cfgv==3.3.1 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.3 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.5 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32" or python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
coverage==6.5.0 ; python_version >= "3.10" and python_version < "4.0"
debugpy==1.6.3 ; python_version >= "3.10" and python_version < "4.0"
dill==0.3.5.1 ; python_version >= "3.10" and python_version < "4.0"
distlib==0.3.6 ; python_version >= "3.10" and python_version < "4.0"
django-debug-toolbar==3.7.0 ; python_version >= "3.10" and python_version < "4.0"
django-extensions==3.2.1 ; python_version >= "3.10" and python_version < "4.0"
django==4.1.2 ; python_version >= "3.10" and python_version < "4.0"
djhtml==1.5.2 ; python_version >= "3.10" and python_version < "4.0"
filelock==3.8.0 ; python_version >= "3.10" and python_version < "4.0"
flake8==5.0.4 ; python_version >= "3.10" and python_version < "4.0"
identify==2.5.6 ; python_version >= "3.10" and python_version < "4.0"
iniconfig==1.1.1 ; python_version >= "3.10" and python_version < "4.0"
isort==5.10.1 ; python_version >= "3.10" and python_version < "4.0"
lazy-object-proxy==1.7.1 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==2.1.1 ; python_version >= "3.10" and python_version < "4.0"
mccabe==0.7.0 ; python_version >= "3.10" and python_version < "4.0"
mypy-extensions==0.4.3 ; python_version >= "3.10" and python_version < "4.0"
nodeenv==1.7.0 ; python_version >= "3.10" and python_version < "4.0"
packaging==21.3 ; python_version >= "3.10" and python_version < "4.0"
pastel==0.2.1 ; python_version >= "3.10" and python_version < "4.0"
pathspec==0.10.1 ; python_version >= "3.10" and python_version < "4.0"
platformdirs==2.5.2 ; python_version >= "3.10" and python_version < "4.0"
pluggy==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
poethepoet==0.16.2 ; python_version >= "3.10" and python_version < "4.0"
pre-commit==2.20.0 ; python_version >= "3.10" and python_version < "4.0"
psycopg2-binary==2.9.4 ; python_version >= "3.10" and python_version < "4.0"
py==1.11.0 ; python_version >= "3.10" and python_version < "4.0"
pycodestyle==2.9.1 ; python_version >= "3.10" and python_version < "4.0"
pyflakes==2.5.0 ; python_version >= "3.10" and python_version < "4.0"
pylint-django==2.5.3 ; python_version >= "3.10" and python_version < "4.0"
pylint-plugin-utils==0.7 ; python_version >= "3.10" and python_version < "4.0"
pylint==2.15.3 ; python_version >= "3.10" and python_version < "4.0"
pyparsing==3.0.9 ; python_version >= "3.10" and python_version < "4.0"
pytest-django==4.5.2 ; python_version >= "3.10" and python_version < "4.0"
pytest==7.1.3 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==0.21.0 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.10" and python_version < "4.0"
setuptools==65.4.1 ; python_version >= "3.10" and python_version < "4.0"
sqlparse==0.4.3 ; python_version >= "3.10" and python_version < "4.0"
toml==0.10.2 ; python_version >= "3.10" and python_version < "4.0"
tomli==2.0.1 ; python_version >= "3.10" and python_version < "4.0"
tomlkit==0.11.5 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2022.4 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
virtualenv==20.16.5 ; python_version >= "3.10" and python_version < "4.0"
werkzeug==2.2.2 ; python_version >= "3.10" and python_version < "4.0"
wrapt==1.14.1 ; python_version >= "3.10" and python_version < "4.0"
